SignIn/SignUp and some Task functionality
